\name{empiricalControlQA}
\alias{empiricalControlQA}
\title{Quality assessment and filtering threshold selection for empirical controls}
\usage{
empiricalControlQA(object, sdCutoff = 0.1)
}
\arguments{
  \item{object}{\code{MethylSet} object}

  \item{sdCutoff}{Standard deviation cut-off for filtering
  empirical controls}
}
\description{
Quality assessment and filtering threshold selection for
empirical controls
}

