\name{empiricalControlCoverage}
\alias{empiricalControlCoverage}
\title{Visualization of empirical control coverage across technical covariates}
\usage{
empiricalControlCoverage(object, sdThreshold = 0.1)
}
\arguments{
  \item{object}{\code{MethylSet} object}

  \item{sdThreshold}{Standard deviation cut-off for
  filtering empirical controls}
}
\description{
Visualization of empirical control coverage across
technical covariates
}

