\name{fresco-package}
\alias{fresco-package}
\alias{fresco}
\docType{package}
\title{
fresco: preprocessing and normalization for illumina 450k methylation array
}
\description{
fresco provides functions for normalizing data using a flexible local regression
surface that is robust to abberant DNA methylation as well as differences in cell
type admixtures.  It also provides functions for quality control of empirical control
probes as well as filtering for CpGs that contain SNPs or autosomal probes that cross hybridize to sex chromosomes.
}
\author{
Paul Manser \email{manserpt@vcu.edu}
Mark Reimers \email{mreimers@vcu.edu}

Maintainer: Paul Manser \email{manserpt@vcu.edu}
}
\references{
Chen,Y. et al. (2013) Discovery of cross-reactive probes and polymorphic CpGs in the Illumina Infinium HumanMethylation450 microarray. Epigenetics, 8, 203–9.
}

